+++ src/Controller/InvoiceController.php
@@ -45,13 +45,21 @@ class InvoiceController extends AbstractController
     }
 
     #[Route('/new', name: 'app_invoices_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, InvoiceRepository $invoiceRepository): Response
+    public function new(Request $request, InvoiceRepository $invoiceRepository, string $defaultInvoiceAccount, AccountRepository $accountRepository): Response
     {
         $invoice = new Invoice();
         $form = $this->createForm(InvoiceNewType::class, $invoice);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            if (!empty($defaultInvoiceAccount)) {
+                $account = $accountRepository->findOneBy(['value' => $defaultInvoiceAccount]);
+
+                if ($account) {
+                    $invoice->setDefaultReceiverAccount($account);
+                }
+            }
+
             $invoice->setRecorded(false);
             $invoice->setTotalPrice(0);
             $invoiceRepository->save($invoice, true);